Split out draw_next_char().

+++ string.h
@@ -3,10 +3,9 @@
     University of Kent at Canterbury.*/
 #ifndef JBXVT_STRING_H
 #define JBXVT_STRING_H
-#include <stdint.h>
 #include <xcb/xcb.h>
 /*  Display the string at the current position.
     new_line_count is the number of new lines in the string.  */
-void jbxvt_string(xcb_connection_t * xc, uint8_t * restrict str, uint_fast16_t
-	len, const int8_t new_line_count);
+void jbxvt_string(xcb_connection_t * xc, uint8_t * restrict str, int len,
+	const int new_line_count);
 #endif//!JBXVT_STRING_H
